+++ TerminusClient/src/edu/buffalo/cse/terminus/client/sensors/SoundAlgo.java
@@ -12,7 +12,7 @@ public class SoundAlgo {
 	
 	private MediaRecorder Recorder = null;
 	public float[] slevel;
-	public static int SOUND_FACTOR=20000;
+	public static float SOUND_FACTOR=20000;
 	public static boolean FirstSndPri = false;
 	public static int samplerate = 1;
 	public static double sound;
